# application/services/custom_services/productlist/src/models/productlist.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const productlistSchema = new Schema({
   created_date: { type: Date, default: Date.now },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date, default: Date.now },
   name: String,
   price: String,
   quantity: String
})

const productlistModel = mongoose.model('productlist', productlistSchema, 'productlist');
export default productlistModel;